requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asn1crypto==1.5.1
bitarray==3.8.0
boto3==1.40.74
botocore==1.40.74
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.1
cryptography==46.0.3
distro==1.9.0
Django==5.2.8
docx==0.2.4
dotenv==0.9.9
filelock==3.20.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
image==1.5.33
jiter==0.12.0
jmespath==1.0.1
lxml==6.0.2
numpy==2.3.4
oauthlib==3.3.1
openai==2.8.0
packaging==25.0
pandas==2.3.3
pdfreader==0.1.15
pillow==12.0.0
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowflake==1.9.0
snowflake-connector-python==4.0.0
snowflake-snowpark-python==1.42.0
snowflake._legacy==1.0.2
snowflake.core==1.9.0
sortedcontainers==2.4.0
sqlparse==0.5.3
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
xlsxwriter==3.2.9
